ispravak prebrisane slike

+++ Home(1)/src/com/example/home1/Text.java
@@ -19,7 +19,7 @@ public class Text {
 	
 	//koji je rating dao trenutno logirani korisnik za ovaj answer
 	//-1 ako nije rateao
-	public float userRating;
+	public int userRating;
 	
 	public Text() {}
 }
